Configs

+++ frontend/src/components/WorkoutDetails.js
@@ -1,30 +1,40 @@
-import { useWorkoutContext } from "../hooks/useWorkoutContext"
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
-const WorkoutDetails = ({ workout }) =>{//With the object destructuring { workout }, it extracts the workout property from the object and assigns it to a variable named workout within the component. This allows you to directly access the workout property without having to use props.workout.
-    const { dispatch } = useWorkoutContext()
+// date fns
+import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
-    const handleClick = async ()=>{
-        const response = await fetch("/api/workouts/"+workout._id,{
-            method: "DELETE"
-        })
+const WorkoutDetails = ({ workout }) => {
+  const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
-        const json = await response.json()
-
-        if (response.ok){
-            dispatch({type: "DELETE_WORKOUT",payload: json})
-        }
-
+  const handleClick = async () => {
+    if (!user) {
+      return
     }
 
-    return (
-        <div className="workout-details">
-            <h4>{workout.title}</h4>
-            <p><strong>Load : </strong>{workout.load}</p>
-            <p><strong>Reps : </strong>{workout.reps}</p>
-            <p>{workout.createdAt}</p>
-            <span onClick={handleClick}>delete</span>
-        </div>
-    )
+    const response = await fetch('/api/workouts/' + workout._id, {
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
+    })
+    const json = await response.json()
+
+    if (response.ok) {
+      dispatch({type: 'DELETE_WORKOUT', payload: json})
+    }
+  }
+
+  return (
+    <div className="workout-details">
+      <h4>{workout.title}</h4>
+      <p><strong>Load (kg): </strong>{workout.load}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+    </div>
+  )
 }
 
 export default WorkoutDetails
